JAMES-1889 Don't bind PostDequeueDecoratorFactory in ActiveMQ guice module
