+++ server/container/guice/guice-common/src/main/java/org/apache/james/modules/server/ActiveMQQueueModule.java
@@ -21,10 +21,8 @@ package org.apache.james.modules.server;
 
 import javax.jms.ConnectionFactory;
 
-import org.apache.james.jmap.send.PostDequeueDecoratorFactory;
 import org.apache.james.queue.activemq.ActiveMQMailQueueFactory;
 import org.apache.james.queue.api.MailQueueFactory;
-import org.apache.james.queue.api.MailQueueItemDecoratorFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -40,7 +38,6 @@ public class ActiveMQQueueModule extends AbstractModule {
     @Override
     protected void configure() {
         bind(EmbeddedActiveMQ.class).in(Scopes.SINGLETON);
-        bind(MailQueueItemDecoratorFactory.class).to(PostDequeueDecoratorFactory.class).in(Scopes.SINGLETON);
     }
     
     @Provides
