+++ server/container/guice/guice-common/src/main/java/org/apache/james/modules/server/RawPostDequeueDecoratorModule.java
@@ -16,22 +16,19 @@
  * specific language governing permissions and limitations      *
  * under the License.                                           *
  ****************************************************************/
-
 package org.apache.james.modules.server;
 
-import org.apache.james.jmap.send.PostDequeueDecoratorFactory;
-import org.apache.james.queue.api.MailQueueFactory;
+import org.apache.james.queue.api.MailQueueItemDecoratorFactory;
+import org.apache.james.queue.api.RawMailQueueItemDecoratorFactory;
 
 import com.google.inject.AbstractModule;
 import com.google.inject.Scopes;
 
-public class MemoryMailQueueModule extends AbstractModule {
+public class RawPostDequeueDecoratorModule extends AbstractModule {
 
     @Override
     protected void configure() {
-        bind(MemoryMailQueueFactory.class).in(Scopes.SINGLETON);
-        bind(PostDequeueDecoratorFactory.class).in(Scopes.SINGLETON);
-
-        bind(MailQueueFactory.class).to(MemoryMailQueueFactory.class);
+        bind(MailQueueItemDecoratorFactory.class).to(RawMailQueueItemDecoratorFactory.class).in(Scopes.SINGLETON);
     }
+
 }
